+++ ui/client/components/toolbars/group/buttons/GroupFinishButton.tsx
@@ -1,11 +1,10 @@
 import React from "react"
-import {useDispatch, useSelector} from "react-redux"
-import {finishGrouping, groupSelected} from "../../../../actions/nk/groups"
-import {CapabilitiesToolbarButton} from "../../../toolbarComponents/CapabilitiesToolbarButton"
-import {getGraph, getGroupingState} from "../../../../reducers/selectors/graph"
 import {useTranslation} from "react-i18next"
+import {useDispatch, useSelector} from "react-redux"
+import {finishGrouping} from "../../../../actions/nk"
 import {ReactComponent as Icon} from "../../../../assets/img/toolbarButtons/group-finish.svg"
-import {canGroupSelection} from "../../../../reducers/graph/utils"
+import {getGroupingState} from "../../../../reducers/selectors/graph"
+import {CapabilitiesToolbarButton} from "../../../toolbarComponents/CapabilitiesToolbarButton"
 
 export function GroupFinishButton(): JSX.Element {
   const groupingState = useSelector(getGroupingState)
@@ -19,22 +18,6 @@ export function GroupFinishButton(): JSX.Element {
       icon={<Icon/>}
       disabled={(groupingState || []).length <= 1}
       onClick={() => dispatch(finishGrouping())}
-    />
-  )
-}
-
-export function GroupSelectedButton(): JSX.Element {
-  const graph = useSelector(getGraph)
-  const {t} = useTranslation()
-  const dispatch = useDispatch()
-
-  return (
-    <CapabilitiesToolbarButton
-      write
-      name={t("panels.actions.group-selected.button", "group selected")}
-      icon={<Icon/>}
-      disabled={!canGroupSelection(graph)}
-      onClick={() => dispatch(groupSelected())}
     />
   )
 }
