+++ ui/client/components/toolbars/group/buttons/UngroupButton.tsx
@@ -1,12 +1,11 @@
 import React from "react"
-import {getSelectedGroups} from "../../../../reducers/graph/utils"
-import {useDispatch, useSelector} from "react-redux"
-import NodeUtils from "../../../graph/NodeUtils"
-import {ungroup, ungroupSelected} from "../../../../actions/nk/groups"
-import {CapabilitiesToolbarButton} from "../../../toolbarComponents/CapabilitiesToolbarButton"
-import {getGraph, getNodeToDisplay} from "../../../../reducers/selectors/graph"
 import {useTranslation} from "react-i18next"
+import {useDispatch, useSelector} from "react-redux"
+import {ungroup} from "../../../../actions/nk/groups"
 import {ReactComponent as Icon} from "../../../../assets/img/toolbarButtons/ungroup.svg"
+import {getNodeToDisplay} from "../../../../reducers/selectors/graph"
+import NodeUtils from "../../../graph/NodeUtils"
+import {CapabilitiesToolbarButton} from "../../../toolbarComponents/CapabilitiesToolbarButton"
 
 function UngroupButton(): JSX.Element {
   const nodeToDisplay = useSelector(getNodeToDisplay)
@@ -20,22 +19,6 @@ function UngroupButton(): JSX.Element {
       icon={<Icon/>}
       disabled={!NodeUtils.nodeIsGroup(nodeToDisplay)}
       onClick={() => dispatch(ungroup(nodeToDisplay))}
-    />
-  )
-}
-
-export function UngroupSelectedButton() {
-  const graph = useSelector(getGraph)
-  const {t} = useTranslation()
-  const dispatch = useDispatch()
-
-  return (
-    <CapabilitiesToolbarButton
-      write
-      name={t("panels.actions.ungroup-selected.button", "ungroup selected")}
-      icon={<Icon/>}
-      disabled={!getSelectedGroups(graph).length}
-      onClick={() => dispatch(ungroupSelected())}
     />
   )
 }
